+++ src/hooks/diary/register/useDiaryRegister.ts
@@ -2,6 +2,7 @@ import {BACKEND_URL} from '@env';
 import {useNavigation} from '@react-navigation/native';
 import {FEELING_STATUS} from '@src/constants/feelingStatus';
 import {SANITIZE_OPTION} from '@src/constants/sanitizeOption';
+import {getNewAccessToken} from '@src/libs/getNewAccessToken';
 import {Diary} from '@src/types/diary';
 import {NavigationProps} from '@src/types/navigation';
 import {useMutation, useQueryClient} from '@tanstack/react-query';
@@ -42,6 +43,11 @@ export const useDiaryRegister = (
         body: JSON.stringify(diaryData),
       });
 
+      if (response.status === 401) {
+        await getNewAccessToken();
+        throw new Error('Access token has expired.');
+      }
+
       if (!response.ok) {
         throw new Error('Failed to register.');
       }
@@ -52,6 +58,7 @@ export const useDiaryRegister = (
       await queryClient.invalidateQueries({queryKey: ['diaryList']});
       navigation.goBack();
     },
+    retry: 1,
     throwOnError: true,
   });
