+++ src/hooks/diary/update/useDiaryUpdate.ts
@@ -2,6 +2,7 @@ import {BACKEND_URL} from '@env';
 import {useNavigation} from '@react-navigation/native';
 import {FEELING_STATUS} from '@src/constants/feelingStatus';
 import {SANITIZE_OPTION} from '@src/constants/sanitizeOption';
+import {getNewAccessToken} from '@src/libs/getNewAccessToken';
 import {Diary} from '@src/types/diary';
 import {NavigationProps} from '@src/types/navigation';
 import {useMutation, useQueryClient} from '@tanstack/react-query';
@@ -50,6 +51,11 @@ export const useDiaryUpdate = (
         body: JSON.stringify(diaryData),
       });
 
+      if (response.status === 401) {
+        await getNewAccessToken();
+        throw new Error('Access token has expired.');
+      }
+
       if (!response.ok) {
         throw new Error('Failed to update.');
       }
@@ -60,6 +66,7 @@ export const useDiaryUpdate = (
       await queryClient.invalidateQueries({queryKey: ['diaryList']});
       navigation.goBack();
     },
+    retry: 1,
     throwOnError: true,
   });
