+++ src/hooks/survey/content/useSurveyContentItemList.ts
@@ -1,4 +1,5 @@
 import {BACKEND_URL} from '@env';
+import {getNewAccessToken} from '@src/libs/getNewAccessToken';
 import {SurveyContentList} from '@src/types/survey';
 import {useSuspenseQuery} from '@tanstack/react-query';
 import EncryptedStorage from 'react-native-encrypted-storage';
@@ -12,11 +13,13 @@ export const useSurveyContentItemList = (
       const accessToken = await EncryptedStorage.getItem('access_token');
       const response = await fetch(
         `${BACKEND_URL}/api/media?type=${contentCategory}&page=0&size=636`,
-        {
-          method: 'GET',
-          headers: {Cookie: `access_token=${accessToken}`},
-        },
+        {method: 'GET', headers: {Cookie: `access_token=${accessToken}`}},
       );
+
+      if (response.status === 401) {
+        await getNewAccessToken();
+        throw new Error('Access token has expired.');
+      }
 
       if (!response.ok) {
         throw new Error('Failed to fetch data.');
@@ -26,7 +29,7 @@ export const useSurveyContentItemList = (
     },
     staleTime: 600000,
     gcTime: 1800000,
-    retry: false,
+    retry: 1,
   });
 
   return {surveyContentList: data};
