+++ src/hooks/tour/useTourItemList.tsx
@@ -1,4 +1,5 @@
 import {BACKEND_URL} from '@env';
+import {getNewAccessToken} from '@src/libs/getNewAccessToken';
 import {SavedPlan} from '@src/types/plan';
 import {useSuspenseQuery} from '@tanstack/react-query';
 import EncryptedStorage from 'react-native-encrypted-storage';
@@ -10,10 +11,13 @@ export const useTourItemList = () => {
       const accessToken = await EncryptedStorage.getItem('access_token');
       const response = await fetch(`${BACKEND_URL}/api/travel/user-plan`, {
         method: 'GET',
-        headers: {
-          Cookie: `access_token=${accessToken}`,
-        },
+        headers: {Cookie: `access_token=${accessToken}`},
       });
+
+      if (response.status === 401) {
+        await getNewAccessToken();
+        throw new Error('Access token has expired.');
+      }
 
       if (!response.ok) {
         return [];
@@ -23,6 +27,7 @@ export const useTourItemList = () => {
     },
     staleTime: 1000 * 60 * 10,
     gcTime: 1000 * 60 * 30,
+    retry: 1,
   });
 
   return {tourItemList: data};
